+++ pkg/mat32/internal/math32/export_test.go
@@ -1,0 +1,5 @@
+package math32
+
+// Export internal functions for testing.
+
+var SqrtGo = sqrt